move incremental add/remove static methods.
